Refactor: Coin 관련 컴포넌트 분리 작업 #103
